requirements.txt
python-deriv-api
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
xgboost>=2.0.0
scikit-learn>=1.3.0
httpx
python-dotenv